feat(roles): [#56049] rework roles code
This commit ensures that permissions in the realm-builder are generate
similarly to how they are generated in OS2mo, and also prepares the
realm-builder for pulling the roles directly from OS2mo, in the future
when OS2mo exposes an APIRoles API.

During this rewrite, it was noticed that the current realm-builder does
in fact not create `delete` and `refresh` roles at all, thus this has
been resolved, and these roles are now created in Keycloak as they are
in OS2mo.
Additionally composite roles are now created using regexes instead of
using a hardcoded list of low-level roles, ensuring that they
automatically pick up appropriate roles as they are added in the future.

A new set of composite roles have been added namely the `x_admin` roles,
which give full access to the entity type `x`, for instance
`address_admin`, which gives the user holding the role full access to
read and write all `address` entities.