feat(client): [#57330] introduce hardcoded uuid client claims
The purpose of the change is to support audit logging in OS2mo, by
ensuring that each client has a unique UUID that can be traced back to
the integration using the client, and thus traced back to the
integration that modified a given data entry in OS2mo.

This commit hardcodes UUIDs for the default DIPEX client and the default
orgviewer client, and allows for a UUID to be set by deployment for
clients created by the integration-builder.

A separate UUID from the client UUID generated by keycloak must be used,
as we consider Keycloak ephemeral and as such consider the UUIDs
generated for clients to be ephemeral by extension. Having the UUIDs be
hardcoded and commited to the repository (as for DIPEX and orgviewer) or
having them be explicitly provided in the deployment mechanism ensures
that they will be consistent across keycloak resets.

The format for the hardcoded UUIDs is:
```
{{hexspeak name}}-baad-c0de-{{ascii description with - seperator}}
```
For DIPEX this becomes:
```
d1fec000-baad-c0de-0000-004449504558
```
Where `d1fec000` is hexspeak for "dipex", and `4449504558` are ascii
codes for "DIPEX".

Simarily for the SD integration:
```
5d05d05d-baad-c0de-0000-53444c6f656e
```
Where `5d05d05d` is hexspeak for "sd sd sd" and `53444c6f656e` are ascii
codes for "SDLoen".