feat(auth)!: [#59555] remove dipex client
Previously the DIPEX client could be constructed with the ordinary
realm-builder, however after this commit it must be constructed using
the integration realm-builder.

This commit is to ensure that each integration makes its own client,
thereby ensuring that we can track which integration makes which change
in MO, w.r.t. the audit log.

BREAKING CHANGE:

The integration builder can no longer build the DIPEX client.
If you need the DIPEX client, please construct it using the integration
builder instead.