feat!: [#61289] don't give owner every permission
BREAKING CHANGE:

Prior to this change, owners were given every other role available in
OS2mo. This is no longer the case. The change should have no impact on
the service-API, as that doesn't implement RBAC anyway (only admin and
owner checks). The intention is to implement proper collection-based
RBAC together with owner in GraphQL; this won't work if the owner role
overrides all other roles.